+++ Examples/coSwiftDemo/coSwiftDemoTests/coSwiftPromiseTests.swift
@@ -73,6 +73,18 @@ func testPromise22() -> Promise<String> {
     }
 }
 
+func promiseAsyncA(parameter: String) -> Promise<String> {
+    return Promise<String>(constructor: { (fulfill, _) in
+        fulfill(parameter + "promiseAsyncA")
+    })
+}
+
+func promiseAsyncB(parameter: String) -> Promise<String> {
+    return Promise<String>(constructor: { (fulfill, _) in
+        fulfill(parameter + "promiseAsyncB")
+    })
+}
+
 class PromiseSpec: QuickSpec {
     
     
@@ -282,6 +294,64 @@ class PromiseSpec: QuickSpec {
                         co.join()
                         expect(step).to(equal(1))
                         done()
+                    }
+                })
+            }
+            
+            it("test chained promise") {
+                
+                promiseAsyncA(parameter: "A")
+                    .then { (msg)  in promiseAsyncB(parameter: msg) }
+                    .then { (promiseB) -> Void in
+                        expect(promiseB).to(equal("ApromiseAsyncApromiseAsyncB"))
+                }
+            }
+            
+            it("test co delay") {
+                
+                let queue = DispatchQueue(label: "queue")
+                var step = 0
+                let co = co_launch(queue: queue) {
+                    
+                    print("time before:\(Date())")
+                    
+                    try co_delay(3)
+                    step = 1
+                    print("time after:\(Date())")
+                }
+                
+                waitUntil(timeout: 5, action: { (done) in
+                    
+                    co_launch {
+                        co.join()
+                        expect(step).to(equal(1))
+                        done()
+                    }
+                })
+            }
+            
+            it("test cancel co delay") {
+                
+                let queue = DispatchQueue(label: "queue")
+                var step = 0
+                let co = co_launch(queue: queue) {
+                    
+                    print("time before:\(Date())")
+                    
+                    try co_delay(3)
+                    step = 1
+                    print("time after:\(Date())")
+                }
+                
+                waitUntil(timeout: 5, action: { (done) in
+                    
+                    co_launch {
+                        co.cancelAndJoin()
+                        
+                        DispatchQueue.main.asyncAfter(deadline: .now() + 3, execute: {
+                            expect(step).to(equal(0))
+                            done()
+                        })
                     }
                 })
             }